43 bug localplace 등록시 column content type id cannot be null 발생

+++ src/main/java/com/example/ieumapi/localplace/dto/LocalPlaceSearchResponse.java
@@ -14,7 +14,7 @@ public record LocalPlaceSearchResponse (
     String userNickName,
     Source source, // 데이터 출처 (USER, KTO)
     String ktoContentId, // 한국관광공사 contentId, nullable
-    Long contentTypeId,
+    int contentTypeId,
     PlaceCategory category,
     String imageUrl,
     String sumNailUrl,
